Add files via upload
This CSS code snippet can be used to create and customize custom cursors for a website, with different sizes and styles based on the screen size of the device. Note that this code alone won't fully enable a custom cursor functionality; it requires some JavaScript or other code to handle the cursor behavior based on user interactions.